requirements.txt
numpy>=1.24.3
pandas>=2.0.0
scikit-learn>=1.2.2
imbalanced-learn>=0.10.1
joblib>=1.2.0


nltk>=3.8.1
spacy>=3.5.2

tensorflow>=2.12.0
transformers>=4.28.1

fastapi>=0.95.1
uvicorn>=0.21.1
python-multipart>=0.0.6
requests>=2.28.2
pydantic>=1.10.7


streamlit>=1.22.0
plotly>=5.14.1
streamlit-option-menu>=0.3.2
streamlit-extras>=0.2.7

matplotlib>=3.7.1
seaborn>=0.12.2

pyyaml>=6.0

-e .